+++ lib/src/loader.dart
@@ -6,11 +6,11 @@ import 'default_theme.dart';
 enum LoaderTheme { dialogDefault, dialogCircle }
 
 class DialogLoader {
-  final ValueNotifier<Widget> _title = ValueNotifier<Widget>(Container());
-  final ValueNotifier<Widget> _leftIcon = ValueNotifier<Widget>(Container());
-
-  final ValueNotifier<Widget> _rightIcon = ValueNotifier<Widget>(Container());
+  final ValueNotifier<Widget> _title = ValueNotifier<Widget>(SizedBox());
+  final ValueNotifier<Widget> _leftIcon = ValueNotifier<Widget>(SizedBox());
+  final ValueNotifier<Widget> _rightIcon = ValueNotifier<Widget>(SizedBox());
   final ValueNotifier<bool> _barrierDismissible = ValueNotifier<bool>(false);
+  final ValueNotifier<double> _avatarRadius = ValueNotifier<double>(25.0);
 
   bool _dialogIsOpen = false;
   late BuildContext _context;
@@ -25,10 +25,12 @@ class DialogLoader {
     Widget? rightIcon,
     bool barrierDismissible: false,
     bool autoClose: true,
+    double? avatarRadius,
   }) {
-    _title.value = title ?? Container();
-    _leftIcon.value = leftIcon ?? Container();
-    _rightIcon.value = rightIcon ?? Container();
+    _title.value = title ?? SizedBox();
+    _leftIcon.value = leftIcon ?? SizedBox();
+    _rightIcon.value = rightIcon ?? SizedBox();
+    _avatarRadius.value = avatarRadius ?? 25.0;
     _barrierDismissible.value = barrierDismissible;
     if (autoClose) _closeAfterUpdate();
   }
@@ -51,20 +53,22 @@ class DialogLoader {
 
   show({
     LoaderTheme theme: LoaderTheme.dialogDefault,
-    Widget? title,
-    Widget? leftIcon,
-    Widget? rightIcon,
+    Widget title: const SizedBox(),
+    Widget leftIcon: const SizedBox(),
+    Widget rightIcon: const SizedBox(),
     Color backgroundColor: Colors.white,
     Color barrierColor: Colors.black26,
     bool barrierDismissible: false,
     double elevation: 5.0,
     double borderRadius: 5.0,
+    double avatarRadius: 25.0,
   }) {
     _dialogIsOpen = true;
-    _title.value = title ?? Container();
+    _title.value = title;
     _barrierDismissible.value = barrierDismissible;
-    _rightIcon.value = rightIcon ?? Container();
-    _leftIcon.value = leftIcon ?? Container();
+    _rightIcon.value = rightIcon;
+    _leftIcon.value = leftIcon;
+    _avatarRadius.value = avatarRadius;
 
     return showGeneralDialog(
       context: _context,
@@ -74,7 +78,13 @@ class DialogLoader {
           children: [
             AnimatedBuilder(
               animation: Listenable.merge(
-                [_title, _leftIcon, _rightIcon, _barrierDismissible],
+                [
+                  _title,
+                  _leftIcon,
+                  _rightIcon,
+                  _barrierDismissible,
+                  _avatarRadius,
+                ],
               ),
               builder: (BuildContext context, _) {
                 return GestureDetector(
@@ -94,6 +104,7 @@ class DialogLoader {
                           context: _context,
                           leftIcon: _leftIcon,
                           title: _title,
+                          avatarRadius: _avatarRadius,
                           rightIcon: _rightIcon,
                           barrierColor: barrierColor,
                           elevation: elevation,