+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [0.0.3+3] - 2021-06-29.
+
+* Add max-width and max-height of the loader container
+* Update the example code
+* Change default widget from Container to Sizebox
+* ~~Add new properti avatarRadius (now you can resize the circle size from the option) to the dialogCirle theme.~~
 ## [0.0.1] - 2021-06-26.
 
 * Add new theme of the loader